refactor(sync): runtime sync status был переведён на прямую проекцию из SyncCoordinator
ICloudSyncStatus вынесен в отдельный тип, а ICloudSyncStatusService удалён как лишний adapter поверх SyncCoordinator
AppComposition, AppDependencies и SettingsScreenController теперь читают runtime sync status напрямую из app-level SyncCoordinator или AppState, без дополнительной status projection
