+++ README.md
@@ -377,7 +377,7 @@
 - [x] сузить app-level remote reload correlation до конкретного `sync-backed` store: текущий reload wiring уже требует pair `NSPersistentStoreRemoteChange + finished import`, но пока не сопоставляет store identity между `remote change` и `CloudKit` event context; нужно выпускать reload только для совпавшего sync-backed store и не реагировать на несвязанные или future local-only/store-adjacent notifications;
 - [x] покрыть store-scoped remote reload correlation дополнительными integration-тестами: отдельно зафиксировать сценарии mismatched `storeIdentifier` / `storeUUID`, local-only store notifications, повторные import events без нового remote change и stale pending state после failure/cancellation, чтобы reload contract не оставался вероятностным;
 - [x] довести app-level логирование и диагностику sync bootstrap/runtime до полного operational contract: отдельно логировать container setup failures, schema/bootstrap failures, bootstrap fallback reasons, account/status resolution, merge/import/export failures и итоговые reload decisions так, чтобы будущая device validation шла по заранее известным log markers без ad-hoc instrumentation;
-- [ ] определить и зафиксировать окончательную projection boundary для runtime sync state: решить, остаётся ли `ICloudSyncStatusService` самостоятельным read-model adapter поверх `SyncCoordinator` или удаляется в пользу прямого app-level runtime source, после чего выровнять чтение sync status между `SyncCoordinator`, `AppState`, `Settings Screen` и shell-level consumers без дублирующих status projections;
+- [x] определить и зафиксировать окончательную projection boundary для runtime sync state: решить, остаётся ли `ICloudSyncStatusService` самостоятельным read-model adapter поверх `SyncCoordinator` или удаляется в пользу прямого app-level runtime source, после чего выровнять чтение sync status между `SyncCoordinator`, `AppState`, `Settings Screen` и shell-level consumers без дублирующих status projections;
 - [ ] провести cleanup / refactor app-level sync orchestration после стабилизации runtime contract: сузить обязанности `AppDependencies` вокруг remote reload/bootstrap wiring, убрать одноразовые diagnostic helpers и оставшиеся временные glue-ветки, а также оставить одну понятную композицию между `AppComposition`, `AppDependencies`, `SyncCoordinator` и runtime sources;
 - [ ] подготовить deferred real-device validation kit для sync: собрать явный checklist сценариев, ожидаемых status/log transitions и минимального smoke-test matrix для пары simulator/device или device/device, чтобы после появления платного Apple Developer аккаунта оставалось выполнить уже предопределённый validation pass, а не заново исследовать поведение синхронизации.
