+++ RSSReader/Infrastructure/AppComposition.swift
@@ -150,12 +150,10 @@ struct AppRootContainer: View {
                 appState.applyInterfaceThemeMode(settings.interfaceThemeMode)
             }
 
-            if let iCloudSyncStatusService = dependencies.iCloudSyncStatusService {
-                let iCloudSyncStatus = try iCloudSyncStatusService.currentStatus()
-                if appState.iCloudSyncStatus != iCloudSyncStatus {
-                    appState.applyICloudSyncStatus(iCloudSyncStatus)
-                }
-            }
+            AppComposition.applyCurrentICloudSyncStatus(
+                from: dependencies.syncCoordinator,
+                to: appState
+            )
 
             if settings.selectedSourcesFilterRawValue != restoredFilter.rawValue {
                 _ = try appSettingsService.updateSettings(
