+++ RSSReader/Services/Sync/ICloudSyncStatus.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+enum ICloudSyncStatus: Equatable, Sendable {
+    case disabled
+    case statusUnavailable
+    case idle
+    case syncing
+    case failed(String)
+}
